Adding tests for the in-memory data store
